renamed reporting.py to metrics.py
