+++ case_studies/galsim_galaxies/get_figures.py
@@ -14,16 +14,16 @@ from matplotlib.figure import Figure
 from torch import Tensor
 from tqdm import tqdm
 
-from bliss import generate, reporting
+from bliss import generate, metrics
 from bliss.catalog import FullCatalog, TileCatalog
 from bliss.datasets.galsim_galaxies import GalsimBlends
 from bliss.datasets.sdss import convert_mag_to_flux
 from bliss.encoder import Encoder
+from bliss.metrics import compute_bin_metrics, get_boostrap_precision_and_recall, match_by_locs
 from bliss.models.decoder import ImageDecoder
 from bliss.models.galaxy_net import OneCenteredGalaxyAE
 from bliss.models.psf_decoder import PSFDecoder
 from bliss.plotting import BlissFigure, plot_image, scatter_shade_plot
-from bliss.reporting import compute_bin_metrics, get_boostrap_precision_and_recall, match_by_locs
 
 ALL_FIGS = ("single_gal", "blend_gal", "toy")
 
@@ -204,11 +204,11 @@ class AutoEncoderReconRandom(BlissFigure):
 
         recon_no_background = recon_means - background.cpu()
         assert torch.all(recon_no_background.sum(axis=(1, 2, 3)) > 0)
-        true_meas = reporting.get_single_galaxy_measurements(
+        true_meas = metrics.get_single_galaxy_measurements(
             noiseless_images, psf_image.reshape(-1, 53, 53), sdss_pixel_scale
         )
         true_meas = {f"true_{k}": v for k, v in true_meas.items()}
-        recon_meas = reporting.get_single_galaxy_measurements(
+        recon_meas = metrics.get_single_galaxy_measurements(
             recon_no_background, psf_image.reshape(-1, 53, 53), sdss_pixel_scale
         )
         recon_meas = {f"recon_{k}": v for k, v in recon_meas.items()}
